Add policies, authorizers in controllers and relations to padalinys

+++ app/Http/Controllers/Admin/BannerController.php
@@ -6,15 +6,11 @@ use App\Models\Banner;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 use App\Http\Controllers\Controller as Controller;
+use App\Http\Controllers\ResourceController;
 use App\Models\Padalinys;
 
-class BannerController extends Controller
+class BannerController extends ResourceController
 {
-
-    public function __construct()
-    {
-        $this->authorizeResource(Banner::class, 'banner');
-    }
 
     /**
      * Display a listing of the resource.
@@ -23,6 +19,9 @@ class BannerController extends Controller
      */
     public function index(Request $request)
     {
+        
+        $this->authorize('viewAny', [Banner::class, $this->authorizer]);
+        
         $title = request()->input('title');
         $padaliniai = request()->input('padaliniai');
 
@@ -51,6 +50,8 @@ class BannerController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', [Banner::class, $this->authorizer]);
+        
         return Inertia::render('Admin/Content/CreateBanner');
     }
 
@@ -62,6 +63,8 @@ class BannerController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', [Banner::class, $this->authorizer]);
+        
         $request->validate([
             'title' => 'required',
             'image_url' => 'required',
@@ -91,6 +94,7 @@ class BannerController extends Controller
     public function show(Banner $banner)
     {
         //
+        $this->authorize('view', [Banner::class, $banner, $this->authorizer]);
     }
 
     /**
@@ -101,6 +105,8 @@ class BannerController extends Controller
      */
     public function edit(Banner $banner)
     {
+        $this->authorize('update', [Banner::class, $banner, $this->authorizer]);
+        
         return Inertia::render('Admin/Content/EditBanner', [
             'banner' => $banner,
         ]);
@@ -115,6 +121,8 @@ class BannerController extends Controller
      */
     public function update(Request $request, Banner $banner)
     {
+        $this->authorize('update', [Banner::class, $banner, $this->authorizer]);
+        
         $banner->title = $request->title;
         $banner->is_active = $request->is_active;
         $banner->link_url = $request->link_url ?? "";
@@ -132,6 +140,8 @@ class BannerController extends Controller
      */
     public function destroy(Banner $banner)
     {
+        $this->authorize('delete', [Banner::class, $banner, $this->authorizer]);
+        
         $banner->delete();
 
         return redirect()->route('banners.index')->with('info', 'Baneris ištrintas!');
